requirements.txt
Flask
flask_session
google-generativeai